Convert ink_hrtime to module
