+++ include/proxy/http/HttpTransact.h
@@ -51,6 +51,8 @@
 #include <string>
 #include <string_view>
 
+import tscore;
+
 #define HTTP_OUR_VIA_MAX_LENGTH 1024 // 512-bytes for hostname+via string, 512-bytes for the debug info
 
 #define HTTP_RELEASE_ASSERT(X) ink_release_assert(X)
