+++ include/tscore/ink_time.h
@@ -37,7 +37,6 @@
 
 #include "tscore/ink_platform.h"
 #include "tscore/ink_defs.h"
-#include "tscore/ink_hrtime.h"
 
 /*===========================================================================*
