+++ src/iocore/cache/CacheDoc.cc
@@ -25,9 +25,9 @@
 
 #include "iocore/eventsystem/IOBuffer.h"
 
-#include "tscore/ink_hrtime.h"
-
 #include <cstring>
+
+import tscore;
 
 namespace
 {
@@ -80,7 +80,7 @@ void
 Doc::pin(std::uint32_t const pin_in_cache)
 {
   // coverity[Y2K38_SAFETY:FALSE]
-  this->pinned = static_cast<uint32_t>(ink_get_hrtime() / HRTIME_SECOND) + pin_in_cache;
+  this->pinned = static_cast<uint32_t>(ink_get_hrtime() / HRTIME_SECONDS(1)) + pin_in_cache;
 }
 
 void
