+++ src/traffic_server/traffic_server.cc
@@ -1990,7 +1990,6 @@ main(int /* argc ATS_UNUSED */, const char **argv)
 #endif
 
   // Pick the system clock to choose, likely only on Linux. See <linux/time.h>.
-  extern int gSystemClock; // 0 == CLOCK_REALTIME, the default
   gSystemClock = RecGetRecordInt("proxy.config.system_clock").value_or(0);
 
   if (!command_flag) { // No need if we are going into command mode.
@@ -2132,9 +2131,9 @@ main(int /* argc ATS_UNUSED */, const char **argv)
   // This means any spawn scheduling must be done before this point.
   eventProcessor.start(num_of_net_threads, stacksize);
 
-  eventProcessor.schedule_every(new SignalContinuation, HRTIME_MSECOND * 500, ET_CALL);
-  eventProcessor.schedule_every(new DiagsLogContinuation, HRTIME_SECOND, ET_TASK);
-  eventProcessor.schedule_every(new MemoryLimit, HRTIME_SECOND * 10, ET_TASK);
+  eventProcessor.schedule_every(new SignalContinuation, HRTIME_MSECONDS(500), ET_CALL);
+  eventProcessor.schedule_every(new DiagsLogContinuation, HRTIME_SECONDS(1), ET_TASK);
+  eventProcessor.schedule_every(new MemoryLimit, HRTIME_SECONDS(10), ET_TASK);
   RecRegisterConfigUpdateCb("proxy.config.dump_mem_info_frequency", init_memory_tracker, nullptr);
   init_memory_tracker(nullptr, RECD_NULL, RecData(), nullptr);
 
@@ -2337,7 +2336,7 @@ main(int /* argc ATS_UNUSED */, const char **argv)
   while (!TSSystemState::is_event_system_shut_down()) {
 #if TS_USE_LINUX_IO_URING == 1
     if (ur->valid()) {
-      ur->submit_and_wait(1 * HRTIME_SECOND);
+      ur->submit_and_wait(HRTIME_SECONDS(1));
     } else {
       sleep(1);
     }
