+++ src/tscore/FrequencyCounter.cc
@@ -21,7 +21,28 @@
   limitations under the License.
  */
 
-#include "tscore/FrequencyCounter.h"
+module;
+
+#include <cstdint>
+
+export module tscore:frequency_counter;
+
+import :hrtime;
+
+export class FrequencyCounter
+{
+public:
+  void     increment(uint16_t amount = 1);
+  uint32_t get_count();
+  virtual ~FrequencyCounter() {}
+
+protected:
+  uint16_t   _count[2]    = {0};
+  ink_hrtime _last_update = 0;
+
+private:
+  virtual ink_hrtime _ink_get_hrtime();
+};
 
 void
 FrequencyCounter::increment(uint16_t amount)
